* MOOSandbox: restored compatibility with pre-C++11 compilers  * MOOSandbox: added parameters for number of runs

+++ projects/MOO/PerceptronTest.h
@@ -54,7 +54,7 @@ public:
 
       ExpressionPtr model;
       FitnessPtr fitness;
-      context.enterScope("Function " + ((string) functionNumber));
+      context.enterScope("Function " + string((int) functionNumber));
       context.enterScope("Learning");
       context.resultCallback("x", 0.0);
       learner->solve(context, problem, storeBestSolverCallback(*(ObjectPtr* )&model, fitness));
